requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-openai
langchain-experimental
langchain-core
langchain-text-splitters
networkx
matplotlib
streamlit